Fixes for Windows build
MSVC makes long 32 bits, which creates some confusion in these calls to
std::max.

+++ mlir/lib/Conversion/AIRLoweringPass.cpp
@@ -535,8 +535,8 @@ AIRChannelInterfaceToAIRRtConversionImpl(OpBuilder builder,
       for (auto o : op_strides.drop_back())
         strides[idx++] = builder.create<arith::IndexCastOp>(
             loc, IntegerType::get(ctx, 64), o);
-    idx = 4 - std::max(thisOp.getSizes().size(),
-                       (unsigned long)thisMemrefType.getRank());
+    idx = 4 -
+          std::max(thisOp.getSizes().size(), (size_t)thisMemrefType.getRank());
     // If sizes field is empty, then infer sizes from memref shape
     if (thisOp.getSizes().empty())
       for (auto d : air::getTensorShape(thisMemrefType))
